+++ cluster/src/main/java/io/scalecube/cluster/election/ElectionProtocol.java
@@ -6,10 +6,11 @@ import io.scalecube.cluster.election.api.ElectionEvent;
 import io.scalecube.cluster.election.api.ElectionService;
 import io.scalecube.cluster.election.api.HeartbeatRequest;
 import io.scalecube.cluster.election.api.HeartbeatResponse;
-import io.scalecube.cluster.election.api.Leader;
 import io.scalecube.cluster.election.api.State;
 import io.scalecube.cluster.election.api.VoteRequest;
 import io.scalecube.cluster.election.api.VoteResponse;
+import io.scalecube.cluster.membership.IdGenerator;
+import io.scalecube.transport.Address;
 import io.scalecube.transport.Message;
 import java.time.Duration;
 import java.util.Collection;
@@ -39,6 +40,50 @@ public class ElectionProtocol implements ElectionService {
 
   private String memberId;
 
+  private static class Protocol {
+
+    private static final String HEARTBEAT = "/heartbeat";
+
+    private static final String VOTE = "/vote";
+
+    public static Mono<Message> FALSE_VOTE =
+        Mono.just(new VoteResponse(false, null)).map(Message::fromData);
+
+    /**
+     * create request message in the context of a given election topic.
+     *
+     * @param sender of this request.
+     * @param topic of this election.
+     * @param action of the request.
+     * @param data to be sent.
+     * @return request message.
+     */
+    public static Message asRequest(Address sender, String topic, String action, Object data) {
+      return Message.builder()
+          .sender(sender)
+          .correlationId(IdGenerator.generateId())
+          .qualifier(topic + action)
+          .data(data)
+          .build();
+    }
+
+    public static Message asHeartbeatRequest(Address sender, String topic, HeartbeatRequest data) {
+      return asRequest(sender, topic, HEARTBEAT, data);
+    }
+
+    public static boolean isHeartbeat(String topic, String value) {
+      return (topic + HEARTBEAT).equalsIgnoreCase(value);
+    }
+
+    public static Message asVoteRequest(Address sender, String topic, VoteRequest data) {
+      return asRequest(sender, topic, VOTE, data);
+    }
+
+    public static boolean isVote(String topic, String value) {
+      return (topic + VOTE).equalsIgnoreCase(value);
+    }
+  }
+
   @Override
   public String id() {
     return this.cluster.member().id();
@@ -53,8 +98,8 @@ public class ElectionProtocol implements ElectionService {
     return stateMachine.currentState();
   }
 
-  public Mono<Leader> leader() {
-    return Mono.just(new Leader(this.memberId, stateMachine.leaderId()));
+  public Mono<String> leader() {
+    return Mono.just(stateMachine.leaderId());
   }
 
   public static class Builder {
