--- cluster/src/main/java/io/scalecube/cluster/election/api/Leader.java
@@ -1,27 +1,0 @@
-package io.scalecube.cluster.election.api;
-
-public class Leader {
-
-  private String memberId;
-  private String leaderId;
-
-  public Leader() {}
-
-  public Leader(String memberId, String leaderId) {
-    this.memberId = memberId;
-    this.leaderId = leaderId;
-  }
-
-  public String memberId() {
-    return memberId;
-  }
-
-  public String leaderId() {
-    return leaderId;
-  }
-
-  @Override
-  public String toString() {
-    return "Leader [memberId=" + memberId + ", leaderId=" + leaderId + "]";
-  }
-}
